+++ src/utils/time.ts
@@ -15,6 +15,10 @@ import {
 export const hrToMillis = ([seconds, nanoseconds]: [number, number]) =>
   Math.round((seconds * 1e3) + (nanoseconds / 1e6))
 
+export const hrToMillisFloat = ([seconds, nanoseconds]: [number, number]) => {
+  return (seconds * 1e3) + (nanoseconds / 1e6)
+}
+
 export const hrToNano = ([seconds, nanoseconds]: [number, number]) =>
   seconds * 1e9 + nanoseconds